math.geometry classes now use float

+++ fx-util/src/main/java/com/dua3/fx/util/FxFontUtil.java
@@ -14,7 +14,7 @@
 
 package com.dua3.fx.util;
 
-import com.dua3.utility.math.Dimension2d;
+import com.dua3.utility.math.geometry.Dimension2d;
 import com.dua3.utility.text.FontUtil;
 
 import javafx.scene.text.Font;
@@ -37,7 +37,7 @@ public class FxFontUtil implements FontUtil<Font> {
     @Override
     public Dimension2d getTextDimension(CharSequence s, com.dua3.utility.text.Font f) {
          var bounds = FxUtil.getTextBounds(s, f);
-         return Dimension2d.of(bounds.getWidth(), bounds.getHeight());
+         return Dimension2d.of((float) bounds.getWidth(), (float) bounds.getHeight());
     }
 
     @Override
